Move project team react code to team client folder
