+++ src/apps/investments/client/projects/team/state.js
@@ -1,0 +1,1 @@
+export const TASK_EDIT_PROJECT_TEAM_MEMBERS = 'TASK_EDIT_PROJECT_TEAM_MEMBERS'
