+++ src/apps/investments/client/projects/team/tasks.js
@@ -1,0 +1,16 @@
+import { apiProxyAxios } from '../../../../../client/components/Task/utils'
+import { validateTeamAdvisersAreUnique } from './validator'
+
+export const updateTeamMembers = ({ teamMembers, id }) => {
+  return new Promise(async (resolve, reject) => {
+    const reasonAdvisersAreNotUnique =
+      validateTeamAdvisersAreUnique(teamMembers)
+    if (reasonAdvisersAreNotUnique) {
+      reject(reasonAdvisersAreNotUnique)
+    } else {
+      const url = `v3/investment/${id}/team-member`
+      const response = await apiProxyAxios.put(url, teamMembers)
+      resolve(response)
+    }
+  })
+}
