+++ src/apps/investments/client/projects/team/validator.js
